Add files via upload
This project is implemented for AES algorithm, also called Advanced Encryption Standard or the Rijndael algorithm. We have executed Advanced Encryption Standard (Round 0 to Round 10, with 128-bit key) in Core Java. Encryption  of AES has been implemented in the project. This project reviews the overall structure of AES and focus particularly on the four steps used in each round of AES: (1) byte substitution, (2) shift rows, (3) mix columns, and (4) add round key. The lookup tables for the byte substitution steps in encryption  and the Key Expansion Algorithms for the 128 bit AES has also been included in the project.